+++ app/components/Spinner.jsx
@@ -1,0 +1,17 @@
+import React, {Component} from "react";
+import styles from '../styles/main.css';
+
+class Spinner extends Component {
+
+    render() {
+        return (
+            <div className={styles.make_modal}>
+                <div className={styles.loading_spinner}>
+                </div>
+            </div>
+        );
+    }
+
+}
+
+export default Spinner;