+++ app/pages/TrendsSensor.jsx
@@ -8,6 +8,7 @@ import Map from '../containers/TrendsMap';
 import TrendsParameters from '../containers/TrendsParameters';
 import TrendsSeasons from '../containers/TrendsSeasons';
 import TrendsRegions from '../containers/TrendsRegions';
+import Spinner from '../components/Spinner';
 import {Grid, Cell, Content, List, Card, CardHeader, CardTitle, CardText} from 'react-mdc-web';
 import styles from '../styles/main.css';
 import trendsStyles from '../styles/trends.css';
@@ -30,11 +31,7 @@ class TrendsSensor extends Component {
 
         let loading_spinner;
         if (this.props.show_spinner === true && this.props.parameter !== '') {
-            loading_spinner = (
-                <div className={styles.make_modal}>
-                    <div className={styles.loading_spinner}> </div>
-                </div>
-            );
+            loading_spinner = (<Spinner/>);
         }
 
         let selections = '';