+++ app/reducers/sensorDetail.js
@@ -15,8 +15,11 @@ const sensorDetail = (state:backendsState = defaultState, action) => {
                 showExplorePopup: true
             });
         case RECEIVE_SENSOR:
+            let sensor_datapoints = Object.assign({}, action.sensor_data);
+            let sensor_sources = Object.assign({}, action.sensor_data);
             return Object.assign({}, state, {
-                datapoints: collectdata(action.sensor_data),
+                datapoints: collect_data(sensor_datapoints),
+                sources: collect_sources(sensor_sources),
                 showExplorePopup: false
             });
         case UPDATE_DETAIL:
@@ -38,35 +41,56 @@ const sensorDetail = (state:backendsState = defaultState, action) => {
     }
 };
 
-//TODO: sum to average
 function groupBy(array, col, value, processed) {
     let r = [], o = {};
+    let copy_array = array;
     array.forEach(function (a) {
         if (!o[a[col]]) {
             o[a[col]] = {};
             o[a[col]][col] = a[col];
-            o[a[col]][value] = 0;
             o[a[col]][processed] = a[processed];
+
+            const same_year_values = copy_array.filter(x => x[col] === a[col]);
+            let count = 0;
+            let values = 0;
+            same_year_values.forEach(function(x) {
+                count += x.count;
+                values += x.count * x.average;
+            });
+            o[a[col]][value] = values/count;
             r.push(o[a[col]]);
         }
-        o[a[col]][value] += +a[value];
-        o[a[col]][processed] += +a[processed];
     });
     return r;
 }
 
-function collectdata(data){
-
+function collect_data(data){
     // Handle no return from API
     if (data.length <=0) {
         return [];
     }
 
-    //console.log(data[0]);
+    let output = [];
     for(let key in data[0].properties){
-        data[0].properties[key] = groupBy(data[0].properties[key], 'label', 'average', getProcessedProperty());
+        output[key] = groupBy(data[0].properties[key], 'label', 'average', getProcessedProperty());
     }
-    return data[0].properties;
+    return output;
+}
+
+function  collect_sources(data) {
+    let output = [];
+    for (let key in data[0].properties) {
+        let sources =[];
+        data[0].properties[key].forEach( x => {
+           x.sources.forEach(source => {
+               if(sources.indexOf(source) === -1) {
+                   sources.push(source);
+               }
+           })
+        });
+        output[key] = sources;
+    }
+    return output;
 }
 
 export default sensorDetail